# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontosyx API",
    "description": "The Semantic Orchestrator — Knowledge Graph Lifecycle Platform.\n\n## Response envelope\n\nEvery successful 2xx response is wrapped in:\n\n```json\n{ \"data\": <T>, \"pagination\": {\"next_cursor\": \"...\"}?, \"meta\": {...}? }\n```\n\nEach handler's `responses(... body = T)` documents the type of `data`, **not** the wire shape. Generated clients should either unwrap `data` in a post-processing step (the approach `web/src/lib/api/client.ts` takes — frontend callers receive `T` directly) or wrap every response type in the envelope shape when consuming the spec from a third-party codegen.\n\nError responses use a separate envelope: `{ \"error\": { \"type\": \"...\", \"message\": \"...\" } }` (see the `ErrorResponse` schema).",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{id}/edit": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "edit_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edit commands generated and optionally applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectEditResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty request or no ontology",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Command validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "LLM timeout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "details": {
            "description": "Optional additional details"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "type": {
            "type": "string",
            "description": "Machine-readable error type (e.g., \"not_found\", \"bad_request\")"
          }
        }
      },
...
      "ProjectEditRequest": {
        "type": "object",
        "required": [
          "revision",
          "user_request"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean",
            "description": "If true, returns generated commands without applying them."
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "user_request": {
            "type": "string",
            "description": "Natural language description of the desired ontology change."
          }
        }
      },
      "ProjectEditResponse": {
        "type": "object",
        "required": [
          "commands",
          "explanation"
        ],
        "properties": {
          "commands": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Generated ontology mutation commands."
          },
          "explanation": {
            "type": "string",
            "description": "LLM explanation of what was changed and why."
          },
          "project": {
            "type": [
              "object",
              "null"
            ],
            "description": "Updated project (null in dry_run mode)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}